requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-filter==24.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-dotenv==0.21.0
django-rest-swagger==2.2.0
drf-yasg==1.21.7
cryptography==43.0.0
